lua test fixes effil as lua library
